Continue splitting.
