+++ src/estimagic/optimization/get_algorithm.py
@@ -7,8 +7,8 @@ import numpy as np
 from estimagic.batch_evaluators import process_batch_evaluator
 from estimagic.logging.database_utilities import (
     list_of_dicts_to_dict_of_lists,
-    update_row,
 )
+from estimagic.logging.write_to_database import update_row
 from estimagic.optimization import ALL_ALGORITHMS
 from estimagic.utilities import propose_alternatives
