+++ src/estimagic/optimization/optimization_logging.py
@@ -1,4 +1,5 @@
-from estimagic.logging.database_utilities import append_row, read_last_rows, update_row
+from estimagic.logging.database_utilities import read_last_rows
+from estimagic.logging.write_to_database import append_row, update_row
 
 
 def log_scheduled_steps_and_get_ids(steps, logging, database):
